+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -55,11 +55,11 @@ import {
   CreateUserCsrPayload,
   InitUserCsrPayload,
   UserCsr,
-  InvitationDataVersion,
+  PeerId as QuietPeerId,
 } from '@quiet/types'
 import { CONFIG_OPTIONS, QUIET_DIR, SERVER_IO_PROVIDER, SOCKS_PROXY_AGENT } from '../const'
 import { Libp2pService } from '../libp2p/libp2p.service'
-import { Libp2pEvents, Libp2pNodeParams, Libp2pPeerInfo } from '../libp2p/libp2p.types'
+import { CreatedLibp2pPeerId, Libp2pEvents, Libp2pNodeParams, Libp2pPeerInfo } from '../libp2p/libp2p.types'
 import { LocalDbService } from '../local-db/local-db.service'
 import { LocalDBKeys } from '../local-db/local-db.types'
 import { RegistrationService } from '../registration/registration.service'
@@ -74,8 +74,9 @@ import { ConfigOptions, GetPorts, ServerIoProviderTypes } from '../types'
 import { ServiceState, TorInitState } from './connections-manager.types'
 import { DateTime } from 'luxon'
 import { createLogger } from '../common/logger'
-import { createFromJSON } from '@libp2p/peer-id-factory'
+import { peerIdFromString } from '@libp2p/peer-id'
 import { PeerId } from '@libp2p/interface'
+import { privateKeyFromRaw } from '@libp2p/crypto/keys'
 import { SigChainService } from '../auth/sigchain.service'
 import { Base58, InviteResult } from '3rd-party/auth/packages/auth/dist'
 import { UserService } from '../auth/services/members/user.service'
@@ -335,11 +336,11 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
 
     await this.closeAllServices({ saveTor: true })
 
-    this.logger.info('Resetting StorageService')
-    await this.storageService.clean()
-
     this.logger.info('Cleaning libp2p datastore')
     await this.libp2pService.libp2pDatastore.clean()
+
+    this.logger.info('Resetting StorageService')
+    await this.storageService.clean()
 
     this.logger.info('Purging data')
     await this.purgeData()
@@ -398,10 +399,10 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     // anything to do with Tor.
     this.logger.info('Getting peer ID')
     const peerId = await createPeerId()
-    const peerIdJson = {
-      id: peerId.toString(),
-      pubKey: uint8ArrayToString(peerId.publicKey!, 'base64pad'),
-      privKey: uint8ArrayToString(peerId.privateKey!, 'base64pad'),
+    const peerIdJson: QuietPeerId = {
+      id: peerId.peerId.toString(),
+      privKey: uint8ArrayToString(peerId.privKey.raw, 'base64'),
+      noiseKey: uint8ArrayToString(peerId.noiseKey, 'base64'),
     }
     this.logger.info(`Created network for peer ${peerId.toString()}. Address: ${hiddenService.onionAddress}`)
 
@@ -771,13 +772,17 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     const onionAddress = await this.spawnTorHiddenService(community.id, identity)
 
     this.logger.info(JSON.stringify(identity.peerId, null, 2))
-    const peerId: PeerId = await createFromJSON(identity.peerId)
-    this.logger.info(peerId.toString())
+    const peerIdData: CreatedLibp2pPeerId = {
+      peerId: peerIdFromString(identity.peerId.id),
+      privKey: privateKeyFromRaw(Buffer.from(identity.peerId.privKey, 'base64')),
+      noiseKey: Buffer.from(identity.peerId.noiseKey, 'base64'),
+    }
+    this.logger.info(peerIdData.peerId.toString())
     const peers = filterValidAddresses(community.peerList ? community.peerList : [])
-    const localAddress = createLibp2pAddress(onionAddress, peerId.toString())
+    const localAddress = createLibp2pAddress(onionAddress, peerIdData.peerId.toString())
 
     const params: Libp2pNodeParams = {
-      peerId,
+      peerId: peerIdData,
       listenAddresses: [this.libp2pService.createLibp2pListenAddress(onionAddress)],
       agent: this.socksProxyAgent,
       localAddress: localAddress,
@@ -826,7 +831,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       this.serverIoProvider.io.emit(SocketActionTypes.PEER_DISCONNECTED, payload)
     })
 
-    await this.storageService.init(peerId)
+    await this.storageService.init(peerIdData.peerId)
     // We can use Nest for dependency injection, but I think since the
     // registration service depends on the storage service being
     // initialized, this is helpful to manually inject the storage