+++ packages/backend/src/nest/ipfs-file-manager/ipfs-file-manager.types.ts
@@ -1,3 +1,5 @@
+import { CID } from 'multiformats'
+
 export enum IpfsFilesManagerEvents {
   // Incoming evetns
   DOWNLOAD_FILE = 'downloadFile',
@@ -16,4 +18,29 @@ export interface FilesData {
   message: {
     id: string
   }
+}
+
+export interface ExportProgress {
+  /**
+   * How many bytes of the file have been read
+   */
+  bytesRead: bigint
+
+  /**
+   * How many bytes of the file will be read - n.b. this may be
+   * smaller than `fileSize` if `offset`/`length` have been
+   * specified
+   */
+  totalBytes: bigint
+
+  /**
+   * The size of the file being read - n.b. this may be
+   * larger than `total` if `offset`/`length` has been
+   * specified
+   */
+  fileSize: bigint
+}
+
+export interface ExportWalk {
+  cid: CID
 }