+++ packages/backend/src/nest/storage/storage.types.ts
@@ -30,7 +30,8 @@ export enum UnixFSEvents {
   GET_BLOCK_PROVIDERS = 'blocks:get:providers:get',
   GET_BLOCK = 'blocks:get:blockstore:get',
   WANT_BLOCK = 'bitswap:want-block:block',
-  DOWNLOAD_BLOCK = 'blocks:get:blockstore:put',
+  DOWNLOAD_BLOCK = 'unixfs:exporter:progress:unixfs:file',
+  PUT_BLOCK = 'blocks:get:blockstore:put',
 }
 
 export interface CsrReplicatedPromiseValues {