+++ packages/desktop/src/renderer/components/Channel/File/FileComponent/FileComponent.tsx
@@ -30,7 +30,7 @@ const FileComponentStyled = styled('div')(({ theme }) => ({
   padding: '16px',
   backgroundColor: theme.palette.colors.white,
   borderRadius: '8px',
-  border: `1px solid ${theme.palette.colors.border01}`,
+  border: `1px solid ${theme.palette.colors.border02}`,
 
   [`& .${classes.icon}`]: {
     minWidth: '40px',
@@ -156,7 +156,7 @@ export const FileComponent: React.FC<FileComponentProps & FileActionsProps> = ({
             variant='indeterminate'
             size={18}
             thickness={4}
-            style={{ position: 'absolute', color: theme.palette.colors.lightGray }}
+            style={{ position: 'absolute', color: theme.palette.colors.purple }}
           />
         )
       case DownloadState.Downloading:
@@ -167,7 +167,7 @@ export const FileComponent: React.FC<FileComponentProps & FileActionsProps> = ({
               size={18}
               thickness={4}
               value={100}
-              style={{ position: 'absolute', color: theme.palette.colors.gray }}
+              style={{ position: 'absolute', color: theme.palette.colors.purple }}
             />
             <CircularProgress
               variant='determinate'
@@ -347,7 +347,7 @@ export const FileComponent: React.FC<FileComponentProps & FileActionsProps> = ({
         <div style={{ display: 'flex', width: 'fit-content' }}>
           <div className={classes.icon}>{renderIcon()}</div>
           <div className={classes.filename}>
-            <Typography variant={'h5'} style={{ lineHeight: '20px' }}>
+            <Typography variant={'h5'} style={{ lineHeight: '20px', color: theme.palette.colors.trueBlack }}>
               {name}
               {ext}
             </Typography>