+++ packages/state-manager/src/sagas/appConnection/connection.master.saga.ts
@@ -1,17 +1,15 @@
-import { all, fork, cancelled, takeEvery } from 'typed-redux-saga'
+import { all, takeEvery, fork, cancelled } from 'typed-redux-saga'
 import { uptimeSaga } from './uptime/uptime.saga'
-
-import { type Socket } from '../../types'
-import { createLogger } from '../../utils/logger'
 import { connectionActions } from './connection.slice'
-import { createInviteSaga } from './invite/createInvite.saga'
+import { createLogger } from '../../utils/logger'
+import { onConnectionProcessInfo } from './onConnectionProcessInfo/onConnectionProcessInfo.saga'
 
 const logger = createLogger('connectionMasterSaga')
 
 export function* connectionMasterSaga(socket: Socket): Generator {
   logger.info('connectionMasterSaga starting')
   try {
-    yield all([fork(uptimeSaga), takeEvery(connectionActions.createInvite.type, createInviteSaga, socket)])
+    yield all([fork(uptimeSaga), takeEvery(connectionActions.onConnectionProcessInfo.type, onConnectionProcessInfo)])
   } finally {
     logger.info('connectionMasterSaga stopping')
     if (yield cancelled()) {