+++ packages/state-manager/src/sagas/users/users.selectors.test.ts
@@ -47,6 +47,8 @@ describe('users selectors', () => {
       },
       peerId: {
         id: aliceCertificateData.peerId,
+        privKey: 'foobar',
+        noiseKey: 'barbaz',
       },
     })